% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\docType{data}
\name{code}
\alias{code}
\title{Unicode Tibble}
\format{tibble with columns
- id: decimal number for the code point (rune)
- rune: hexadecimal representation of the rune prefixed with U+
- description: description
- block : block
- countries: comma separated list of countries using the character
- languages: comma separated list of languages using the character
- type: type of rune}
\source{
https://github.com/unicode-table/unicode-table-data
}
\usage{
code
}
\description{
Unicode Tibble
}
\keyword{datasets}
